:recycle: Refactor : apiに認証を追加

+++ pages/api/posts/delete-entry.ts
@@ -1,26 +1,34 @@
 import { NextApiHandler } from 'next'
 import { query } from '../../../lib/db'
+import { getSession } from 'next-auth/client'
 
 const handler: NextApiHandler = async (req, res) => {
+  const session = await getSession({ req })
   const { id } = req.query
-  try {
-    if (!id) {
-      return res.status(400).json({ message: '`id` required' })
+
+  if(session){
+    try {
+      if (!id) {
+        return res.status(400).json({ message: '`id` required' })
+      }
+      if (typeof parseInt(id.toString()) !== 'number') {
+        return res.status(400).json({ message: '`id` must be a number' })
+      }
+      const results = await query(
+        `
+        DELETE FROM entries
+        WHERE id = ?
+        `,
+        id
+      )
+      res.json(results)
+    } catch (e) {
+      res.status(500).json({ message: e.message })
     }
-    if (typeof parseInt(id.toString()) !== 'number') {
-      return res.status(400).json({ message: '`id` must be a number' })
-    }
-    const results = await query(
-      `
-      DELETE FROM entries
-      WHERE id = ?
-      `,
-      id
-    )
-    res.json(results)
-  } catch (e) {
-    res.status(500).json({ message: e.message })
+  } else {
+    res.status(401)
   }
+  res.end()
 }
 
 export default handler
